Changed copy code so it works like FreeDOS Xcopy changed license to match FreeDOS

+++ src/logger.cpp
@@ -3,11 +3,11 @@
 #include <iostream.h>
 #include <time.h>
 #include <string.h>
-#include <iomanip.h>
 
+// Changed to use longs instead of doubles
 void Logger::logTransferDetails(const char* source, const char* destination, 
-                               long fileSize, double maxSpeed, double minSpeed, 
-                               double avgSpeed, double duration) {
+                               long fileSize, long maxSpeed, long minSpeed, 
+                               long avgSpeed, long duration) {
     // Simpler log file handling to avoid errors
     char logPath[256] = "TRANSFER.LOG";  // Use current directory
     
@@ -24,20 +24,20 @@ void Logger::logTransferDetails(const char* source, const char* destination,
     char timeBuffer[80];
     strftime(timeBuffer, 80, "%Y-%m-%d %H:%M:%S", localTime);
 
-    // Convert speeds from bytes/sec to KB/sec (not MB) to avoid large fp numbers
-    double maxSpeedKB = maxSpeed / 1024.0;
-    double minSpeedKB = minSpeed / 1024.0;
-    double avgSpeedKB = avgSpeed / 1024.0;
+    // Convert speeds to KB/sec using integer division
+    long maxSpeedKB = maxSpeed / 1024;
+    long minSpeedKB = minSpeed / 1024;
+    long avgSpeedKB = avgSpeed / 1024;
 
     logFile << "Transfer Log" << endl;
     logFile << "Date and Time: " << timeBuffer << endl;
     logFile << "Source: " << source << endl;
     logFile << "Destination: " << destination << endl;
     logFile << "Size: " << fileSize << " bytes" << endl;
-    logFile << "Max: " << (long)maxSpeedKB << " KB/sec" << endl;
-    logFile << "Min: " << (long)minSpeedKB << " KB/sec" << endl;
-    logFile << "Avg: " << (long)avgSpeedKB << " KB/sec" << endl;
-    logFile << "Time: " << (long)duration << " seconds" << endl;
+    logFile << "Max: " << maxSpeedKB << " KB/sec" << endl;
+    logFile << "Min: " << minSpeedKB << " KB/sec" << endl;
+    logFile << "Avg: " << avgSpeedKB << " KB/sec" << endl;
+    logFile << "Time: " << duration << " seconds" << endl;
     logFile << "----------------------------------------" << endl;
 
     logFile.close();
