+++ src/logger.h
@@ -3,9 +3,10 @@
 
 class Logger {
 public:
+    // Changed to use longs instead of doubles
     void logTransferDetails(const char* source, const char* destination, 
-                            long fileSize, double maxSpeed, double minSpeed, 
-                            double avgSpeed, double duration);
+                            long fileSize, long maxSpeed, long minSpeed, 
+                            long avgSpeed, long duration);
 };
 
 #endif // LOGGER_H
